+++ lib/src/build_mode/build_mode_web.dart
@@ -1,0 +1,5 @@
+/// Run command as `env "dart.vm.product=true" dart test`
+bool isReleaseMode() => const bool.fromEnvironment('dart.vm.product');
+
+/// Run command as `env "dart.vm.profile=true" dart test`
+bool isProfileMode() => const bool.fromEnvironment('dart.vm.profile');