+++ lib/src/constants.dart
@@ -1,3 +1,0 @@
-const bool kReleaseMode = bool.fromEnvironment('dart.vm.product');
-const bool kProfileMode = bool.fromEnvironment('dart.vm.profile');
-const bool kDebugMode = !kReleaseMode && !kProfileMode;