+++ lib/src/dart_test/dart_test.dart
@@ -1,4 +1,4 @@
-import 'dart_test_non_web.dart' if (dart.library.html) 'dart_test_web.dart'
+import 'dart_test_io.dart' if (dart.library.js_interop) 'dart_test_web.dart'
     as platform;
 
 bool isDartTest() => platform.isDartTest();